广义表优化doxygen注释

+++ 5_ArrayAndGList/GenList/src/test.c
@@ -9,15 +9,21 @@
  */
 
 
-#include "test.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include "test.h"
 
 
 /*!
- * @brief 广义表深度测试
+ * @brief **广义表深度测试**
  * @note
+ * 广义表深度测试
+ * ------------
+ * ------------
+ * ###1 构造广义表###
+ * &emsp; 使用字符串"(A,(),(B,(C)))", 调用GenLIstCreateByStr构造广义表
+ * ###2 求广义表的深度###
  */
 void TestGenListDepth() {
     printf("\n");
@@ -25,17 +31,17 @@ void TestGenListDepth() {
     printf("                      Test GenListDepth                      \n");
     printf("                        测试广义表的深度                        \n\n\n");
 
-    ///1 构造广义表###
-    ///&emsp; 使用字符串"(A,(),(B,(C)))", 调用GenLIstCreateByStr构造广义表
+    // ----- 1 构造广义表 -----
+    // 使用字符串"(A,(),(B,(C)))", 调用GenLIstCreateByStr构造广义表
     gen_list_t gen_list = NULL;
 
     char input_gen_list_str[100] = "(A,(),(B,(C)))";
-    GenListCreateByStr(&gen_list, input_gen_list_str, strlen(input_gen_list_str));
+    GenListCreateByStr(&gen_list, input_gen_list_str, (int)strlen(input_gen_list_str));
 
     char* gen_list_str = (char*)malloc(100 * sizeof(char));
     GenListToString(gen_list, gen_list_str, 100);
 
-    ///2 求广义表的深度###
+    // 2 求广义表的深度
     int depth = GenListDepthRecursive(gen_list);
 
     printf("广义表%s的深度: %d\n", gen_list_str, depth);
@@ -45,8 +51,15 @@ void TestGenListDepth() {
 
 
 /*!
- * @brief 广义表复制测试
+ * @brief **广义表复制测试**
  * @note
+ * 广义表复制测试
+ * ------------
+ * ------------
+ * ###1 构造广义表###
+ * &emsp; 使用字符串"(A,(),(B,(C)))", 调用GenLIstCreateByStr构造广义表
+ * ###2 复制广义表###
+ * &emsp; 调用CopyGenList将gen_list复制到new_gen_list
  */
 void TestGenListCopy() {
     printf("\n");
@@ -54,18 +67,18 @@ void TestGenListCopy() {
     printf("                       Test GenListCopy                      \n");
     printf("                        测试广义表的复制                        \n\n\n");
 
-    ///###1 构造广义表###
-    ///&emsp; 使用字符串"(A,(),(B,(C)))", 调用GenLIstCreateByStr构造广义表
+    // ----- 1 构造广义表 -----
+    // 使用字符串"(A,(),(B,(C)))", 调用GenLIstCreateByStr构造广义表
     gen_list_t gen_list = NULL;
     char input_gen_list_str[100] = "(A,(),(B,(C)))";
-    GenListCreateByStr(&gen_list, input_gen_list_str, strlen(input_gen_list_str));
+    GenListCreateByStr(&gen_list, input_gen_list_str, (int)strlen(input_gen_list_str));
 
     char* gen_list_str = (char*)malloc(100 * sizeof(char));
     GenListToString(gen_list, gen_list_str, 100);
     printf("广义表: %s\n\n", gen_list_str);
 
-    /// ###2 复制广义表###
-    /// &emsp; 调用CopyGenList将gen_list复制到new_gen_list
+    // ----- 2 复制广义表 -----
+    // 调用CopyGenList将gen_list复制到new_gen_list
     gen_list_t new_gen_list = NULL;
     GenListCopyRecursive(&new_gen_list, gen_list);
